이미지가 있으면 업데이트

+++ 003 Code/backend/user-service/src/main/java/com/wcd/userservice/service/user/UserServiceImpl.java
@@ -43,11 +43,13 @@ public class UserServiceImpl implements UserService{
 
         String profileImageUrl = null;
 
-        try {
-            fileStore.deleteFile(user.getProfileImage());
-            profileImageUrl = fileStore.storeFile(requestUpdateUser.getProfileImage());
-        } catch (IOException e) {
-            e.printStackTrace();
+        if(requestUpdateUser.getProfileImage() != null && !requestUpdateUser.getProfileImage().isEmpty()) {
+            try {
+                fileStore.deleteFile(user.getProfileImage());
+                profileImageUrl = fileStore.storeFile(requestUpdateUser.getProfileImage());
+            } catch (IOException e) {
+                log.error("Failed to update profile image for user with ID: {}", userId, e);
+            }
         }
 
         user.updateUser(requestUpdateUser, profileImageUrl);
